pyproject.toml
[tool.poetry]
name = "temp-fastapi"
version = "0.1.0"
description = ""
authors = ["bamjun <khy5116@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
apscheduler = "^3.10.4"
uvicorn = "^0.31.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
